clippy/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Ethan Brown <ethan@ewbrowntech.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
